DQN and DQN_CUDA implemented and working
Everything inside DQN_CUDA is supposed to be running on CUDA but idk why CUDA is much slower than the normal one. With nvidia-smi I have seen that CUDA process uses part of the GPU (around 16%) and that does not happen with the normal version, so it should really be running there. Maybe is just a really small problem for CUDA? My next steps are forgetting about CUDA, implementing DoubleDQN, DuelingDQN and DoubleDoubleDQN agents. Maybe after that, I try to implement noisy net again, but better